requirements.txt
Flask==2.3.3
twilio==8.10.0
openai==1.3.7
redis==5.0.1
requests==2.31.0
python-dotenv==1.0.0
speechrecognition==3.10.0
pydub==0.25.1
langdetect==1.0.9
textblob==0.17.1
deep-translator==1.11.4     # Instead of googletrans==4.0.0rc1
whisper==1.1.10
librosa==0.10.1
numpy==1.24.3
scikit-learn==1.3.2
flask-cors==4.0.0
gunicorn==21.2.0